requirements.txt
Django==4.2.2
pyyaml==6.0.1
